Derive the Messenger-Proxy URL from the homeserver URL
As every Messenger-Service needs its own Messenger-Proxy the URL for the Messenger-Proxy cannot be configured at the Webclient because the Webclient is homeserver agnostic.
The configuration of the Webclient still needs a Messenger-Proxy configured to be able to check federation membership of the homeserver prior to login. This functionality should be moved to the registry service in future.

Change-Id: Id1108a51b3e9d4e9a08a22966a078fb7ddd6c5f3